arsac_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
pyjwt==2.8.0
python-dotenv==1.0.1
pydantic==2.7.1
